+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
   - TODO: disable buttons on localhost
 - Allow images as background of section pages
   - Don't actually add the image in CSS, we will do that using inline CSS for now
+- Add a link to the admin backend to the NOFO
 
 ### Changed